forecast screen updated

+++ lib/models/hourly_weather.dart
@@ -196,9 +196,8 @@ class Rain {
   final double threeHours;
 
   const Rain({required this.threeHours});
-
   factory Rain.fromJson(Map<String, dynamic> json) {
-    return Rain(threeHours: json['3h'] ?? 0.0);
+    return Rain(threeHours: (json['3h'] ?? 0).toDouble());
   }
 }
